Move find_fits_keyword to fits_support
